Extract the running the monitor.sh to its own file
