+++ lib/flight_job/models/job.rb
@@ -89,95 +89,6 @@ module FlightJob
       }
     })
 
-    # We have multiple schemas for the monitor response to workaround issues
-    # with JSONSchemer and error reporting on `oneOf` matchers.
-    MONITOR_RESPONSE_SCHEMAS = {
-      initial: JSONSchemer.schema({
-        "type" => "object",
-        "additionalProperties" => true,
-        "required" => ["version", "state"],
-        "properties" => {
-          "version" => { "const" => 1 },
-          "state" => { "enum" => STATES }
-        }
-      }),
-
-      "PENDING" => JSONSchemer.schema({
-        "type" => "object",
-        "additionalProperties" => false,
-        "required" => [],
-        "properties" => {
-          "version" => {}, "state" => {},
-          "scheduler_state" => { "type" => "string", "minLength": 1 },
-          "reason" => { "type" => ["string", "null"] },
-          "start_time" => { "type" => "null" },
-          "end_time" => { "type" => "null" },
-          "estimated_start_time" => { "type" => ["string", "null"] },
-          "estimated_end_time" => { "type" => ["string", "null"] }
-        }
-      }),
-
-      "RUNNING" => JSONSchemer.schema({
-        "type" => "object",
-        "additionalProperties" => false,
-        "required" => ["start_time"],
-        "properties" => {
-          "version" => {}, "state" => {},
-          "scheduler_state" => { "type" => "string", "minLength": 1 },
-          "reason" => { "type" => ["string", "null"] },
-          "start_time" => { "type" => "string", "minLength": 1 },
-          "end_time" => { "type" => "null" },
-          "estimated_start_time" => { "type" => "null" },
-          "estimated_end_time" => { "type" => ["string", "null"] }
-        }
-      }),
-
-      "COMPLETED" => JSONSchemer.schema({
-        "type" => "object",
-        "additionalProperties" => false,
-        "required" => ["start_time"],
-        "properties" => {
-          "version" => {}, "state" => {},
-          "scheduler_state" => { "type" => "string", "minLength": 1 },
-          "reason" => { "type" => ["string", "null"] },
-          "start_time" => { "type" => "string", "minLength": 1 },
-          "end_time" => { "type" => "string", "minLength": 1 },
-          "estimated_start_time" => { "type" => "null" },
-          "estimated_end_time" => { "type" => "null" }
-        }
-      }),
-
-      "CANCELLED" => JSONSchemer.schema({
-        "type" => "object",
-        "additionalProperties" => false,
-        "required" => ["start_time"],
-        "properties" => {
-          "version" => {}, "state" => {},
-          "scheduler_state" => { "type" => "string", "minLength": 1 },
-          "reason" => { "type" => ["string", "null"] },
-          "start_time" => { "type" => ["null", "string"] },
-          "end_time" => { "type" => "string", "minLength": 1 },
-          "estimated_start_time" => { "type" => "null" },
-          "estimated_end_time" => { "type" => "null" }
-        }
-      }),
-
-      "UNKNOWN" => JSONSchemer.schema({
-        "type" => "object",
-        "additionalProperties" => false,
-        "required" => ["start_time"],
-        "properties" => {
-          "version" => {}, "state" => {},
-          "scheduler_state" => { "type" => "string", "minLength": 1 },
-          "reason" => { "type" => ["string", "null"] },
-          "start_time" => { "type" => "null" },
-          "end_time" => { "type" => "null" },
-          "estimated_start_time" => { "type" => "null" },
-          "estimated_end_time" => { "type" => "null" }
-        }
-      })
-    }.tap { |h| h["FAILED"] = h["COMPLETED"] }
-
     def self.load_all
       Dir.glob(new(id: '*').metadata_path).map do |path|
         id = File.basename(File.dirname(path))
@@ -390,62 +301,11 @@ module FlightJob
     end
 
     def submit
-      JobTransitions::SubmitTransition.new(self).run
+      JobTransitions::SubmitTransition.new(self).run!
     end
 
     def monitor
-      # Skip jobs that have terminated, this allows the method to be called liberally
-      if STATES_LOOKUP[state] == :terminal
-        FlightJob.logger.debug "Skipping monitor for terminated job: #{id}"
-        return
-      end
-
-      # Jobs without a scheduler ID should not be in a running/pending state. It is
-      # an error condition if they are
-      unless scheduler_id
-        FlightJob.logger.error "Can not monitor job '#{id}' as it did not report its scheduler_id"
-        metadata['reason'] = "Did not report it's scheduler ID"
-        metadata['state'] = "FAILED"
-        File.write(metadata_path, YAML.dump(metadata))
-        return
-      end
-
-      FlightJob.logger.info("Monitoring Job: #{id}")
-      cmd = [FlightJob.config.monitor_script_path, scheduler_id]
-      execute_command(*cmd, tag: 'monitor') do |status, stdout, stderr, data|
-        if status.success?
-          # Validate the output
-          validate_data(MONITOR_RESPONSE_SCHEMAS[:initial], data, tag: "monitor (initial)")
-          validate_data(MONITOR_RESPONSE_SCHEMAS[data['state']], data, tag: "monitor (#{data['state']})")
-
-          data.each do |key, value|
-            # Ignore the metadata version
-            next if key == "version"
-
-            # Treat empty string/nil as the same value
-            value = nil if value == ''
-
-            # Parse and set times
-            if /_time\Z/.match? key
-              metadata[key] = parse_time(value, type: key)
-
-            # Set other keys
-            else
-              metadata[key] = value
-            end
-          end
-
-          if data['reason'] == ''
-            metadata['reason'] = nil
-          elsif data['reason']
-            metadata['reason'] = data['reason']
-          end
-          File.write(metadata_path, YAML.dump(metadata))
-
-          # Remove the indexing file in terminal state
-          FileUtils.rm_f active_index_path
-        end
-      end
+      JobTransitions::MonitorSingletonTransition.new(self).run!
     end
 
     def decorate
@@ -465,6 +325,10 @@ module FlightJob
       @job_dir ||= File.join(FlightJob.config.jobs_dir, id)
     end
 
+    def terminal?
+      STATES_LOOKUP[state] == :terminal
+    end
+
     protected
 
     def <=>(other)
@@ -473,68 +337,6 @@ module FlightJob
       else
         Time.parse(created_at) <=> Time.parse(other.created_at)
       end
-    end
-
-    private
-
-    def parse_time(time, type:)
-      return nil if ['', nil].include?(time)
-      Time.parse(time).strftime("%Y-%m-%dT%T%:z")
-    rescue ArgumentError
-      FlightJob.logger.error "Failed to parse #{type}: #{time}"
-      FlightJob.logger.debug $!.full_message
-      raise_command_error
-    end
-
-    def validate_data(schema, data, tag:)
-      errors = schema.validate(data).to_a
-      unless errors.empty?
-        FlightJob.logger.error("Invalid #{tag} response for job: #{id}")
-        FlightJob.logger.debug(JSON.pretty_generate(errors))
-        raise_command_error
-      end
-    end
-
-    def execute_command(*cmd, tag:)
-      # NOTE: Should the PATH be configurable instead of inherited from the environment?
-      # This could lead to differences when executed via the CLI or the webapp
-      env = ENV.slice('PATH', 'HOME', 'USER', 'LOGNAME').tap do |h|
-        h['CONTROLS_DIR'] = controls_dir.path
-      end
-      cmd_stdout, cmd_stderr, status = Open3.capture3(env, *cmd, unsetenv_others: true, close_others: true)
-
-      unless status.success?
-        FlightJob.logger.error("Failed to #{tag} job: #{id}")
-      end
-
-      FlightJob.logger.debug <<~DEBUG
-        COMMAND: #{cmd.join(" ")}
-        STATUS: #{status.exitstatus}
-        STDOUT:
-        #{cmd_stdout}
-        STDERR:
-        #{cmd_stderr}
-      DEBUG
-
-      data = nil
-      if status.success?
-        begin
-          data = JSON.parse(cmd_stdout.split("\n").last.to_s)
-        rescue JSON::ParserError
-          FlightJob.logger.error("Failed to parse #{tag} JSON for job: #{id}")
-          FlightJob.logger.debug($!.message)
-          raise_command_error
-        end
-      end
-
-      yield(status, cmd_stdout, cmd_stderr, data)
-    end
-
-    def raise_command_error
-      raise CommandError, <<~ERROR.chomp
-        An error occurred when integrating with the external scheduler service!
-        Please contact your system administrator for further assistance.
-      ERROR
     end
   end
 end
